AngularJs, images and arrow functionality implemented,
revised AngularJs and added in the framework and base functionality. the
arrows to cycle through comics are also working. However they lack any
guard conditions.

I still need to manipulate the url to be able to share specific comics
as currently it will always default to the latest available comic